types: add SwipeInstance type
