+++ src/address-edit/README.zh-CN.md
@@ -122,9 +122,9 @@ export default {
 
 通过 ref 可以获取到 AddressEdit 实例并调用实例方法，详见[组件实例方法](#/zh-CN/advanced-usage#zu-jian-shi-li-fang-fa)。
 
-| 方法名           | 说明         | 参数                  | 返回值 |
-| ---------------- | ------------ | --------------------- | ------ |
-| setAddressDetail | 设置详细地址 | addressDetail: string | -      |
+| 方法名           | 说明         | 参数                    | 返回值 |
+| ---------------- | ------------ | ----------------------- | ------ |
+| setAddressDetail | 设置详细地址 | _addressDetail: string_ | -      |
 
 ### AddressInfo 数据格式
