+++ src/checkbox/index.ts
@@ -5,4 +5,4 @@ const Checkbox = withInstall<typeof _Checkbox>(_Checkbox);
 
 export default Checkbox;
 export { Checkbox };
-export type { CheckboxInstance } from './Checkbox';
+export type { CheckboxInstance } from './types';
