+++ src/form/Form.tsx
@@ -1,9 +1,4 @@
-import {
-  PropType,
-  defineComponent,
-  ExtractPropTypes,
-  ComponentPublicInstance,
-} from 'vue';
+import { PropType, defineComponent, ExtractPropTypes } from 'vue';
 
 // Utils
 import { truthProp, createNamespace } from '../utils';
@@ -19,6 +14,7 @@ import type {
   FieldValidateError,
   FieldValidateTrigger,
 } from '../field/types';
+import type { FormExpose } from './types';
 
 const [name, bem] = createNamespace('form');
 
@@ -41,23 +37,7 @@ const props = {
   },
 };
 
-type FormProps = ExtractPropTypes<typeof props>;
-
-type FormExpose = {
-  submit: () => void;
-  validate: (name?: string | string[] | undefined) => Promise<void>;
-  scrollToField: (
-    name: string,
-    options?: boolean | ScrollIntoViewOptions | undefined
-  ) => void;
-  resetValidation: (name?: string | string[] | undefined) => void;
-};
-
-export type FormInstance = ComponentPublicInstance<FormProps, FormExpose>;
-
-export type FormProvide = {
-  props: FormProps;
-};
+export type FormProps = ExtractPropTypes<typeof props>;
 
 export default defineComponent({
   name,
