+++ src/form/index.ts
@@ -5,4 +5,4 @@ const Form = withInstall<typeof _Form>(_Form);
 
 export default Form;
 export { Form };
-export type { FormInstance } from './Form';
+export type { FormInstance } from './types';
